Adjusting tickets styles
